Some minor changes in the structure.
To respect the npm recommendations: https://docs.npmjs.com/files/package.json#directories
