+++ index.js
@@ -1,0 +1,5 @@
+const bruteforce = require('./lib/bruteforce');
+
+module.exports = {
+  bruteforce
+};
