+++ lib/bruteforce.js
@@ -63,7 +63,7 @@ function filled_array(len, val) {
 	return rv;
 }
 
-bruteforce = function(opts, callback) {
+function bruteforce(opts, callback) {
 	b = new Brute(opts.charset, opts.min, opts.max);
 	b.interval = setInterval(function (){
 		value = b.get_value()
@@ -71,6 +71,4 @@ bruteforce = function(opts, callback) {
 	}, 0);
 };
 
-exports.bruteforce = bruteforce;
-
-
+module.exports = bruteforce;
